third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-parser-only.html

<!DOCTYPE html>
<html>
<head>
<script>
proto = Object.create(HTMLElement.prototype, { createdCallback: { value: function() { this.innerHTML = "Hello"; } } });
document.register("x-foo", { prototype: proto });
</script>
</head>
<body>
<x-foo></x-foo>
</body>
</html>
