src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/lifecycle-created-parser-only-expected.txt

CONSOLE ERROR: line 6: Uncaught TypeError: document.registerElement is not a function
